Add CRNN+CTC OCR, CTC decoder & runtime tools
Introduce a CRNN + CTC OCR pipeline and supporting runtime/config changes across Python training code and Android inference.

Key changes:
- Add CRNN-CTC model and training flow (model_ai/src/ocr/model.py, train.py): build CRNN, CTC training wrapper, dataset helpers, CTC decoding, and save_vocabulary entries for decoder/blank_index.
- Add mobile CTC decoder and vocabulary handling (android/.../CtcTextDecoder.kt, ModelAssetSource, TrainedModelOcrEngine) and update copy_tflite_to_android.ps1 to include vocabulary asset.
- Add unified TensorFlow runtime helpers and dataset builder (model_ai/src/training_runtime.py) and update detection/legibility/trainers to use runtime, mixed precision, and tf.data pipelines.
- Enhance TrOCR training & data loaders (trocr_train.py, trocr_data.py): automatic device selection, AMP/mixed precision, gradient clipping, pin/persistent worker options, and richer training runtime metadata.
- Add GPU check script and small infra improvements (model_ai/scripts/check_gpu.py, setup_venv.ps1, requirements.txt) and update docs/guides (README.md, TROCR_GUIDE.md) to document workflow and options.

These changes add CTC-capable inference/training, improve GPU/mixed-precision handling, and wire vocabulary metadata between training, evaluation, and Android assets.

+++ model_ai/src/detection/model.py
@@ -43,7 +43,7 @@ def build_detector_model(
     u2 = tf.keras.layers.Concatenate()([u2, c1])
     c4 = conv_block(u2, 32)
 
-    outputs = tf.keras.layers.Conv2D(1, 1, activation="sigmoid")(c4)
+    outputs = tf.keras.layers.Conv2D(1, 1, activation="sigmoid", dtype="float32")(c4)
     model = tf.keras.Model(inputs=inputs, outputs=outputs, name="text_region_detector")
     model.compile(
         optimizer=tf.keras.optimizers.Adam(learning_rate=learning_rate),
